Fix firestore path & refactor
* just a small fix & a quick refactor while I was there :-)

- added link to Notion in README
- added profile mode launch config
- renamed widgets MyApp & NewsItemCard to MainPage & CuratableItemCard
- extracted StreamBuilder/ListView from MyApp into CuratablesView
