+++ .vscode/launch.json
@@ -5,9 +5,22 @@
   "version": "0.2.0",
   "configurations": [
     {
-      "name": "curatable",
+      "name": "current device",
       "request": "launch",
       "type": "dart"
+    },
+    {
+      "name": "profile: current device",
+      "request": "launch",
+      "type": "dart",
+      "flutterMode": "profile",
+    },
+    {
+      "name": "profile: mac",
+      "request": "launch",
+      "type": "dart",
+      "flutterMode": "profile",
+      "deviceId": "macos"
     },
     {
       "name": "backend",
