+++ README.md
@@ -1,3 +1,5 @@
 # news_curation
 
 Curating news from noisy feeds.
+
+[Notes](https://adventures-in.notion.site/Curatable-a43fffacad9c4574ad450147436cbeb2)
