+++ lib/widgets/main_page.dart
@@ -1,0 +1,53 @@
+import 'package:firebase_core/firebase_core.dart';
+import 'package:flutter/material.dart';
+import 'package:news_curation/widgets/curatables_view.dart';
+
+class MainPage extends StatefulWidget {
+  const MainPage({Key? key}) : super(key: key);
+
+  @override
+  State<MainPage> createState() => _MainPageState();
+}
+
+class _MainPageState extends State<MainPage> {
+  // Set default `_initialized` and `_error` state to false
+  bool _initialized = false;
+  dynamic _error;
+
+  // Define an async function to initialize FlutterFire
+  void initializeFlutterFire() async {
+    try {
+      // Wait for Firebase to initialize and set `_initialized` state to true
+      await Firebase.initializeApp();
+      setState(() {
+        _initialized = true;
+      });
+    } catch (e) {
+      // Set `_error` state to true if Firebase initialization fails
+      setState(() {
+        _error = e;
+      });
+    }
+  }
+
+  @override
+  void initState() {
+    initializeFlutterFire();
+    super.initState();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    // Show error message if initialization failed
+    if (_error != null) {
+      print(_error);
+    }
+
+    // Show a loader until FlutterFire is initialized
+    if (!_initialized) {
+      return const CircularProgressIndicator();
+    }
+
+    return const CuratablesView();
+  }
+}
