+++ pubspec.yaml
@@ -11,22 +11,22 @@ dependencies:
   flutter:
     sdk: flutter
 
-  url_launcher: any
-  firebase_core: any
-  cloud_firestore: any
+  url_launcher:
+  firebase_core:
+  cloud_firestore:
   
-  fast_immutable_collections: any
-  freezed_annotation: any
-  json_annotation: any
+  fast_immutable_collections:
+  freezed_annotation:
+  json_annotation:
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  build_runner: ^2.0.0
-  flutter_lints: ^1.0.0
-  freezed: any
-  json_serializable: any
+  build_runner:
+  flutter_lints:
+  freezed:
+  json_serializable:
 
 flutter:
   uses-material-design: true
