+++ test/widget_test.dart
@@ -7,12 +7,12 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:news_curation/widgets/my_app.dart';
+import 'package:news_curation/widgets/main_page.dart';
 
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(const MyApp());
+    await tester.pumpWidget(const MainPage());
 
     // Verify that our counter starts at 0.
     expect(find.text('0'), findsOneWidget);
